+++ src/package/themes/default/components/button-group.ts
@@ -1,0 +1,5 @@
+export default {
+    gap: '0.5rem',
+    borderRadius: 'var(--vf-button-border-radius)',
+    disabledOpacity: 'var(--vf-states-disabled-opacity)',
+};